servicebus: splitting out the migration code
