+++ azurerm/internal/services/servicebus/migration/servicebus_namespace_resource_test.go
@@ -1,11 +1,10 @@
-package servicebus_test
+package migration
 
 import (
 	"reflect"
 	"testing"
 
 	"github.com/hashicorp/terraform-plugin-sdk/terraform"
-	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/servicebus"
 )
 
 func TestAccAzureRMServiceBusNamespaceMigrateState(t *testing.T) {
@@ -104,7 +103,7 @@ func TestAccAzureRMServiceBusNamespaceMigrateState(t *testing.T) {
 			ID:         tc.ID,
 			Attributes: tc.Attributes,
 		}
-		is, err := servicebus.ResourceAzureRMServiceBusNamespaceMigrateState(tc.StateVersion, is, nil)
+		is, err := ServiceBusNamespaceResourceMigrateState(tc.StateVersion, is, nil)
 		if err != nil {
 			t.Fatalf("bad: %q, err: %#v", tn, err)
 		}
